improve

+++ src/include/raygfx.h
@@ -99,7 +99,7 @@ extern "C"
 
     /* Thick Line */
     RAYGFXAPI int thickLineRGBA(Sint16 x1, Sint16 y1, Sint16 x2, Sint16 y2,
-                                Uint8 width, Uint8 r, Uint8 g, Uint8 b, Uint8 a);
+                                float width, Uint8 r, Uint8 g, Uint8 b, Uint8 a);
 
     /* Circle */
